Common TCK tests for batch query error paths
